pyproject.toml
[tool.poetry]
name = "xian-py"
version = "0.1.0"
description = "Python tools to interact with the Xian blockchain"
authors = ["endogen"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyNaCl = "^1.5.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
